Refactor layout imports and enhance layout components for consistency
- Updated import paths for layout components to use absolute paths for improved clarity and maintainability.
- Introduced new layout components for chat and non-chat scenarios, enhancing the modularity of the application.
- Adjusted the width of the aside element in RoundsAPage for better layout consistency.
