+++ apps/web/src/app/about/layout.tsx
@@ -1,4 +1,4 @@
-import LayoutWithoutChat from "@/app/layouts/layout-without-chat";
+import LayoutWithoutChat from "@/layouts/layout-without-chat";
 
 export default function AboutLayout({
   children,
