+++ apps/web/src/app/losers/layout.tsx
@@ -1,4 +1,4 @@
-import LayoutWithoutChat from "@/app/layouts/layout-without-chat";
+import LayoutWithoutChat from "@/layouts/layout-without-chat";
 
 export default function LosersLayout({
   children,
