+++ apps/web/src/layouts/layout-with-chat.tsx
