+++ apps/web/src/layouts/layout-without-chat.tsx
